proto: move DnsResponseStream out of dns_response module
